Adds avocado to game
This pull request adds an avocado fruit to the game. It's a fruit, that has potassium and cooking oil in it, medocre yield, above average lfespan, and has a property shared only with bungofruit, that upon eating it completely, it spawns it's pit in it's place.

Growing sprites are borrowed from the lime tree, the avocado, it's pit and the seeds are custom-made sprites instead.

Accidentally reverted, so re-commiting.